[rpc-test] fix flaky TestCancellation
KUDU-3042 added a test case for the interaction between rpc
cancellation and rpc timeouts. This test was flaky as it assumed that
rpcs that are cancelled will return with an aborted status, which is
not always the case due to cancellation being best effort.

The fix is to make the test more permissive by allowing the rpcs to
fail with either an aborted or a timed out status.

Change-Id: If140a9215c94d781c2c61ec30b18d2d67d80310a
Reviewed-on: http://gerrit.cloudera.org:8080/15174
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>