[cmake_modules] shared libs for yaml and gumbo on macOS
I noticed that even in DEBUG build yaml and gumbo libraries were
linked in statically when building on macOS.  It turned out the
system could not find shared libraries since their pattern contained
non-macOS '.so' suffix.

This patch address the issue.  I verified that non-debug binaries
are linked dynamically with the libraries mentioned above.

I also added some extra code to protect against using static libraries
in place of the shared ones.  It's possible to apply this approach
to other thirdparty libraries used by Kudu, but I think it's better
to do so in a separate changelist.

Change-Id: Ie1e56b9dd988501af010bbd1a256671b0460f526
Reviewed-on: http://gerrit.cloudera.org:8080/15182
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>