rpc: allow a sidecar to append multiple slices
This adds support for sidecars to use vectors of Slices instead of a
single Slice each. Currently, this isn't used. However, a later patch
will make use of this to support returning scan batches in a columnar
layout. The row-wise layout could also make use of this to improve
allocator efficiency -- a TODO is added for that.

Note: this changes the Transfer code to use a
boost::container::small_vector instead of a hard-coded std::array with
10 elements. This should have the same "embedded in the object"
performance for responses with a small number of slices, while allowing
an unlimited number of slices when needed.

Change-Id: I05b4d5e5243735db943e315268d837f662891b1a
Reviewed-on: http://gerrit.cloudera.org:8080/15221
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>