[tests] add time anomalies test
This patch adds a new test to exercise the behavior of Kudu masters
and tablet servers in case of various time anomalies.

As of now, a single scenario is present: starting up tablet servers
in case if current time is far behind timestamps in replica's WAL.

This patch also contains a small update on MiniChronyd::SetTime()
to make it more robust if calling the method multiple times
consecutively.

Change-Id: I369f0d78997824324426943a75c126d35d81e264
Reviewed-on: http://gerrit.cloudera.org:8080/15239
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>