[mini-cluster] built-in NTP client configuration mode
This patch introduces configuration mode for the built-in NTP client
in the scope of Kudu external mini-cluster.

Prior to this patch, the only way to configure the built-in NTP client
in a Kudu external mini-cluster was using all available test NTP servers
as reference servers.  With this patch, it's now possible to point
different tablet servers and masters to different NTP servers run
as a part of a test mini-cluster.

A follow-up patch will use the newly introduced mode for a new test.

Change-Id: I88561f757678b664b810b54a27317cac99297403
Reviewed-on: http://gerrit.cloudera.org:8080/15238
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>