KUDU-1563. Add a feature flag for IGNORE operations
This patch adds a master server feature flag to indicate that the cluster
supports `IGNORE` operations. This includes INSERT_IGNORE, DELETE_IGNORE,
and UPDATE_IGNORE. Though this is technically a tserver feature, it is
unreasonable to check if every tablet server supports this feature.
Instead we use the master as a proxy.

In the future KUDU-3211 will add more complete cluster feature flag support.

Additionally this patch leverages the feature flag in the Java client on a
PingRequest to implement a `supportsIgnoreOperations()` method. This
functionality will be used in follow on patches to add ignore operation support
to the Spark and Backup integrations in a compatible way.

Change-Id: I329bd8bde73d247240ae597b677e2cc20a92343a
Reviewed-on: http://gerrit.cloudera.org:8080/16698
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>