[subprocess] use a CountDownLatch
As follow-up to 20be5d957ab80c7c0ed8eaf62324ac70940d7543, this converts
the atomic boolean that signals a shutdown in the SubprocessServer to a
CountDownLatch. This allows us to combine some of the sleeps with the
shutdown checks.

To do this, I had to update the CountDownLatch::CountDown() methods to
return whether they actually counted down.

Change-Id: I1adfdf3a9ccb6e8ee8c60cfab68562dde3e642d6
Reviewed-on: http://gerrit.cloudera.org:8080/15270
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>