[clock] safer handling of metrics-related endpoints
Re-worked the way of retrieving hybrid clock metrics:
  * in case of 'system' time source, do not crash if clock is not
    synchronized when serving /metrics Web server endpoint
  * in general, register function gauges only after HybridClock
    is initialized

In addition, this patch contains a new test scenario to exercise the
control path mentioned above.

This patch also contains a sub-scenario to check for the presence of
metrics introduced in the context of KUDU-3048.

This is a follow-up for 8808b041c, 6371ffdb5, and 7d9d7009a.

Change-Id: Ic172635a7c7b019e5f4394c6107345820d408a68
Reviewed-on: http://gerrit.cloudera.org:8080/15273
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>