mvcc: remove "safe time" terminology
The MvccManager maintains an exclusive lower bound on the timestamps
allowed for assignment for new transactions. It refers to this as the
"safe time".

MVCC's usage of "safe time" is confusing because it is easily conflated
with the very similar "safe time" in the context of consensus: the
timestamp before which the leader promises not to assign a timestamp for
a new write. This is a distributed property of a tablet, managed by the
leader's consensus::TimeManager, whereas MVCC safe time is managed by
each replica.

This replaces MVCC 'safe_time_' and associated methods with the more
literal 'new_txn_timestamp_exc_lower_bound_', in hopes of making it
clearer 1) what "safe time" is by reducing the number of versions of the
term, and 2) how MVCC's "safe time" (now referred to as the "new
transaction timestamp lower bound") is used.

There are no functional changes in this patch.

Change-Id: I990478e20039a6ea541c1aa32a2a865f8e017b83
Reviewed-on: http://gerrit.cloudera.org:8080/15277
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>