[ksck] report on misconfiguration for flag categories
This patch adds functionality to report on misconfiguration of flags
in pre-defined flag categories.  The 'misconfiguration' means different
values for same flags in each group of masters/tablet servers in a
cluster (i.e. intra-group differences), and also differences between
masters' and tablet servers' flags (i.e. cross-group differences).

With this patch, two categories are defined and ready for use:
  * 'time_source' includes the following flags
      ** --builtin_ntp_servers
      ** --time_source
  * 'unusual' includes all experimental, hidden, and unsafe flags

The newly introduced functionality is enabled by default for the
'time_source' flags category.  The list of group categories to check
can be customized by adding
--flags_categories_to_check=<comma-separated-list-of-flags-categories>
to the 'kudu cluster ksck' invocation. Use an empty string to specify
an empty list of flags categories to check.

For testing, I ran a small Kudu cluster and verified that ksck reports
on misconfiguration as expected:
  * if tablet servers are run with different flags
  * if masters are run with flag settings different from those used for
    tablet servers

This patch also contains assorted re-factoring in related areas
of code.  I'm planning to add automated tests to cover the newly
introduced functionality in a separate patch.

Change-Id: I2afbfac9327f85b212ecf8d8f43a2139f90db6bb
Reviewed-on: http://gerrit.cloudera.org:8080/15298
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>