[thirdparty] bump cmake version up to 3.16.4
As it turned out, cmake versions prior to 3.11.2 had issues while
building Kudu with OpenJDK 11: cmake expects the 'javah' binary to be
present in JDK, but that's not the case since JDK 10.  This patch
updates cmake in thirdparty up to 3.16.4 -- the latest released version
as of today.

NOTE: those issues do not manifest itself if JDK of older version
      (e.g., JDK 1.8) is installed at the build machine along with
      JDK 10 or newer.

There was a compilation issue while compiling cmake tests under CentOS6
with devtoolset-3, but Adar provided a work-around recipe to disable
building and running cmake tests.  So, this patch also disables
building/running cmake tests, avoiding the compilation breakage and
saving time and CPU resources.

NOTE: Adar filed the following issue for cmake upstream:
        https://gitlab.kitware.com/cmake/cmake/issues/20401

Change-Id: I1596f6127e3fc6d5715220bb39f1a893c8a8a8ae
Reviewed-on: http://gerrit.cloudera.org:8080/15317
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins