[scripts] Update parse_metrics_log.py to make it usable in other scripts
This wraps the functionality in parse_metrics_log.py in a class
so that it can be used from another python script.

It maintains the old functionality by moving the print statement
to the main function, from the process function.

I will use this so I can wrap this script in some of my performance
analysis work.

Change-Id: Ia107b54a7efc45f9877e7c0b9405cefb550c38f2
Reviewed-on: http://gerrit.cloudera.org:8080/15336
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins