[hms] Fix testAlterTableHandler in later Hive versions
This patch adjusts testAlterTableHandler to stop reusing
already created table objects when creating new tables.
This prevents issues if read-only properties are set.

An example of a read-only property is the ID field added in
HIVE-20556.

The error seen was:
`testAlterTableHandler(org.apache.kudu.hive.metastore.TestKuduMetastorePlugin)InvalidObjectException(message:Id shouldn't be set but table default.table has the Id set to 1. It's a read-only option)`

Change-Id: I8f8ddbe30a3e51957cafd8f1b596745b3262ca36
Reviewed-on: http://gerrit.cloudera.org:8080/15349
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Grant Henke <granthenke@apache.org>