[subprocess] Fix subprocess JAR path
f4176b1a was tested before c35871df was merged in. While they
didn't conflict code-wise, the kudu-subprocess-echo.jar that f4176b1a
tries to run in a test doesn't exist anymore since c35871df.

This commit fixes the location of the JAR and provides the main class as
it's not in the manifest of the JAR anymore.

Change-Id: Icce4088580fcccfe34bddafb0d84f3835091f27c
Reviewed-on: http://gerrit.cloudera.org:8080/15358
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins