KUDU-2972 Add Ranger client
Adds a Ranger client that communicates with the Java subprocess that
wraps the Ranger plugin. It can be used by an authorization provider to
translate between Kudu and Ranger.

Removes default_database out parameter from ParseRangerTableIdentifier
in table_util as we decided to take a slightly different path and this
information is not needed anymore.

Current test coverage consists of only unit tests with a mock subprocess
server. Real integration tests will be added in a follow-up patch once a
working MiniRanger is added to facilitate this testing.

Change-Id: Ie2e1ec19ed3aeb4d82ad38fe1fb655f57021c1a4
Reviewed-on: http://gerrit.cloudera.org:8080/15206
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>