subprocess: report metrics even when calls fail
This moves metric reporting for metrics returned by the subprocess into
the SubprocessServer so they are still reported in the event that the
call fails.

To do this, I moved the SubprocessMetrics into the SubprocessServer,
which means that these metrics are now required by the server. As such,
I moved the declarations of Echo metrics out of subprocess_proxy-test so
they could be reused by other tests.

This should prove more useful, particularly given the fact that failed
calls warrant a deeper look at these metrics.

Change-Id: I87cc82f82a0a767383284799ec6159d680526ea9
Reviewed-on: http://gerrit.cloudera.org:8080/15366
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>