[docker] Fix docker build
This patch fixes the docker build by moving the `java` directory
`COPY` above the C++ build. This is neccessary now that the
subprocessor depends on the Java build.

I also added a missing label, `vcs-ref`, back to the kudu-python
image and made the kudu-python image a default target given
we want to publish it with each release.

Change-Id: Ifbf32dc0c803b3a1ab44724d45dd1e64728d6133
Reviewed-on: http://gerrit.cloudera.org:8080/15388
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>