[docker] Add an Impala quickstart image
This patch adds an `impala` target that can be built and
published for use in quickstart examples. The image is a
single node Impala cluster with a Hive metastore backed
by Derby. This image is purpose build for quickstart usage
only and shouldn't be used for production or multi-node
deployments. The final image size is ~2GB uncompressed.

The build and runtime images reuse the dev and runtime base
images from Kudu given the additional libraries needed are
nominal in size and the reuse cuts back on code duplication.

This image will be used in an Impala quickstart guide that
will be added in a follow on patch.

Currently the image only works on Ubuntu distributions due
to Impala side script assumptions. Follow on patches will
fix these issues for other distributions.

Change-Id: I2d88bcb52f793c0fc83c4f84a9b56c5b961e58d5
Reviewed-on: http://gerrit.cloudera.org:8080/15393
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins