subprocess: maintain a thread for fork/exec
If SubprocessServer::Init() is called from a short-lived thread, because
it does a fork/exec, the subprocess will silently be reaped when the
thread exits.

Specifically, we saw this happening upon initializing the
CatalogManager, which happens in a short-lived thread, when starting up
the Ranger client. This surfaced as us getting an EOF from the
subprocess receiver thread.

Change-Id: I803b1613ef1a988df1da4c908c2c37e1fbbdcf81
Reviewed-on: http://gerrit.cloudera.org:8080/15398
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>