[build] Fail the LINT build with Java static analysis issues
This patch changes the Gradle build to fail when static analysis
issues are found. In the past warnings would be logged and reports
were created, but the build would still succeed.

A `-PignoreCheckFailures` was added in case users wanted to
manually revert back to the old behavior.

Additionally the LINT Jenkins build was updated to run the
static analysis checks and fail if any are found.

Change-Id: I6fb8494639b1bbf9be6d2bd214a986b101496525
Reviewed-on: http://gerrit.cloudera.org:8080/15409
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>