[docker] Use ccache and Gradle cache across builds
This patch uses buildkits `--mount=type=cache` syntax to allow
the builds ccache and Gradle cache to be used across Docker builds.

This speeds up the Docker build when re-building the thirdparty,
C++, and Java modules and mirrors the best practices for
non-Docker based builds.

See the buildkit documentation for details:
https://github.com/moby/buildkit/blob/master/frontend/dockerfile/docs/experimental.md#run---mounttypecache

Below shows the runtime difference for subsequent builds of the
thirdparty target and Kudu target before and after this change:

`thirdparty` target:
- Before: 1:18:13
- After:    0:17:47

`kudu` target:
- Before: 0:19:10
- After:    0:04:08

Change-Id: Ie21ed47983b990e9d2aac419454b9a37a0600334
Reviewed-on: http://gerrit.cloudera.org:8080/16181
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Reviewed-by: Greg Solovyev <gsolovyev@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>