[gradle] remove failOnNoMatchingTests
I found it confusing when gradle reports success on executing test
targets which don't exist.  Prior to this patch, the following returned
success and claimed successfully executing tasks:

  $ ./gradlew :kudu-client:test --tests org.apache.kudu.client.Nonsense

  BUILD SUCCESSFUL in 1s
  15 actionable tasks: 1 executed, 14 up-to-date

This patch fixes it by removing failOnNoMatchingTests.

It seems there were some ramifications on the pattern matching for
dist-test some time ago, but I verified prior and after this patch
and saw that the following works and breaks the same way:

works:
  ../build-support/dist_test.py java loop -n 8 '*ITClient'
  ../build-support/dist_test.py java loop -n 8 '*ITClient*'
  ../build-support/dist_test.py java loop -n 8 org/apache/kudu/client/TestConnectToCluster

doesn't work:
  ../build-support/dist_test.py java loop -n 8 '*ITClient.class'
  ../build-support/dist_test.py java loop -n 8 '*ITClient*X'
  ../build-support/dist_test.py java loop -n 8 org/apache/kudu/client/TestConnectToCluster.class

Change-Id: Iabb0f72fb9481c775ebcc9881a39180fb0413154
Reviewed-on: http://gerrit.cloudera.org:8080/15428
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>