[subprocess] copy subprocess JAR to be next to the master binary
This patch copies the subprocess JAR to be next to the master/tserver
binary instead of linking. It can reduce the coupling from the Java
build and the C++ build, since the Java build could remove the linked
jar if running C++ build at the same time. The same for the hms JAR as
we have similar expectation of the location of the JAR.

Change-Id: I0bab628fa2d1fbb8d9edee32b480c30da2756cdc
Reviewed-on: http://gerrit.cloudera.org:8080/15438
Tested-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>