[build] Fix buildSrc lock timeout
I have seen a few instances where the `make` build fails in a fresh
environment with the below error about a buildSrc lock:

`Timeout waiting to lock buildSrc build lock. It is currently in use by another Gradle instance.`

However, I have not seen this at all in pre-commit builds where
the environment isn’t wiped or in developer environments.

I think the issue is that building the buildSrc module for the first time
takes too long because it needs to pull dependencies and doesn’t
have a cache to use.

This patch adds an extra step in the cmake Gradle initialization
to actually call a project task to ensure the buildSrc module is
initialized.

Change-Id: I26444e6d2f8fffca1d070a591710bd7382063624
Reviewed-on: http://gerrit.cloudera.org:8080/15470
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins