metrics: replace kudu::Bind usage with lambdas in FunctionGauge
Let's start with a little history lesson. Back in 2014, we imported
Chromium's Bind (and associated machinery) into gutil as kudu::Bind. At the
time we hadn't migrated to C++11, so we couldn't use lambdas, and the
kudu::Bind feature set was rich[1], at least as compared to boost::bind.

Now that we're on C++11 and can use lambdas, there's very little that
kudu::Bind can do that lambdas can't. The only thing that comes to mind is
passing ownership of an object to a callback via the kudu::Owned() or
kudu::Passed() wrappers, and we're not using those features. So in the
interest of modernizing the Kudu codebase a bit, I'm going to replace
kudu::Bind with lambdas. Lambdas are simpler, can be more easily inlined by
the compiler, and will result in a reduction of code at the end when we
remove kudu::Bind and friends from the repo.

This patch kicks off the conversion process by using lambdas in
FunctionGauge, which is a heavily used source of kudu::Bind.

1. See commit 66282df07.

Change-Id: If60a0dabde1fde1465d73dd9e8edc244fc780c37
Reviewed-on: http://gerrit.cloudera.org:8080/15474
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>