[build] Skip builds with no functional changes
This patch introduces the option to skip builds in pre-commit when
the changes are in files or directories that do not impact the build
or test.

This functionality is disabled by default but can be enabled on our
Jenkins instance by setting `KUDU_ALLOW_SKIPPED_TESTS=1`.

I also hid the `DONT_BUILD` functionality behind the
`KUDU_ALLOW_SKIPPED_TESTS` flag to ensure commits with
`DONT_BUILD` in them do not prevent other builds that are using
this script from running.

Change-Id: If56db6593ce7aa64b562a9fb277a337bffc7b532
Reviewed-on: http://gerrit.cloudera.org:8080/15481
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>