[util] KUDU-3067 add OpenStack metadata detector
This patch adds OpenStack metadata detector that works with OpenStack
Nova metadata server (see [1] for details).  In addition, this patch
fixes the existing AWS detector to tell apart a true EC2 instance
from a masquerading OpenStack one [2].

I couldn't get access to an OpenStack instance, but I asked the reporter
of KUDU-3067 to test how it works and report back.

1. https://docs.openstack.org/nova/latest/user/metadata.html#metadata-service
2. https://docs.openstack.org/nova/latest/user/metadata.html#metadata-ec2-format

Change-Id: I84cc6d155ab1fbd7b401f5349d292f46fcac3a34
Reviewed-on: http://gerrit.cloudera.org:8080/15488
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: liusheng <liusheng2048@gmail.com>