[util] GetBindIpForDaemon() consumes PID wider than 18 bit
This patch updates the logic of GetBindIpForDaemon() to accept process
identifiers wider than 18 bits.  Prior to this patch, maximum allowable
PID was 262143, and feeding in a higher PID lead to a crash.  The
limitation was due to the mapping scheme of (PID, index) --> loopback
IP address: it mapped 18-bit PID and 6-bit index into the 24-bit space
of 127.0.0.0/8 loopback addresses.

With this patch, GetBindIpForDaemon() stops imposing the limit on the
maximum PID.  Now the uniqueness of the generated IP address is not
guaranteed in general if OS generates PIDs wider than 18 bits, but it
still holds across all Kudu test processes if both of the following
conditions are true:

  * PIDs are monotonically increasing with a delta D which is much
    smaller than 262143 for newly spawned processes (usually D is 1)

  * no more than 262144/D processes are spawned between the earliest
    still running process and the latest running one while executing
    Kudu tests involving external mini-cluster

The former item is not true for hardened Linux distros where PID
assignment is randomized, but it's unlikely that anybody is interested
running Kudu tests on such systems.

Essentially, the new scheme uses PID % 262144 as input for the PID part,
keeping it in the 18-bit space even if the original PID address space is
wider (it might be up to 22 bits wide for 64-bit Linux distros).

This patch helps running Kudu tests on Docker containers where the host
OS is configured to have PIDs higher than 262143 and updating maximum
PID is not an option since /proc/sys/kernel/pid_max is mounted
read-only.

This patch relates to the following JIRA issues:
  KUDU-1334
  KUDU-3044

Change-Id: Idf2162a2a976a0ce997f02a321f230e7f5878328
Reviewed-on: http://gerrit.cloudera.org:8080/15519
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins