[docker] Make the build image more useful for development
This patch makes the docker build image more useful for
development. It contains a few small changes including:
- Fixed Sentry in thirdparty
- Added lsof package to fix HMS tests
- Replaced ntp with chrony
- Added build packages (cmake, ninja-build)
- Added vim for basic file editing
- Added ccache. A follow on change will allow the docker cache to
share the ccache between builds too.

With this patch most of the tests pass in the Docker container.
The following tests are still failing:
- hybrid_clock-test
- client-stress-test
- memory_gc-itest
- raft_consensus-itest.2

Fixes for these will be added in follow on patches.

I used the following steps to build a xenial image and run the tests:

   export TARGETS="build"
   ./docker/docker-build.sh
   docker run -it --rm apache/kudu:build-latest /bin/bash
   cd /kudu
   mkdir build/debug
   cd build/debug
   ../../thirdparty/installed/common/bin/cmake \
     -DCMAKE_BUILD_TYPE=debug ../..
   make -j6
   ctest -j6

Change-Id: I1e4b463a440a3f76fb07425aebcbda2050b9917e
Reviewed-on: http://gerrit.cloudera.org:8080/15530
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>