rpc: relax limit on number of sidecars
Following the example of 0a7787f6563ae316a4a6207a9244c19ba8a7731f this
changes the the sidecars for outbound calls to use a
boost::container::small_vector instead of a fixed size array. This means
we no longer need such a strict limit.

This commit raises the limit to 10000 as a reasonable "can't see the
need for more than this" number.

This also fixes an issue where the underlying socket syscalls fail if
the number of slices in the iovec list is more than IOV_MAX. The new
test verifies the behavior.

Change-Id: I8e53d5b50753e4aa57d885718ad5b24558636f82
Reviewed-on: http://gerrit.cloudera.org:8080/15547
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>