[clock] hybrid_clock-test to run in different environments
This patch updates two scenarios of the hybrid_clock-test to make
the suite pass if run under:
  * containerized environments where ntp_adjtime() requires
    extra privileges (e.g., Docker prior to v 18.06.0-ce)
  * environments where the local clock is not NTP-synchronized

For the TimeSourceAutoSelection scenario, it's about adding
pre-conditions for those two particular cases.  If the pre-conditions
are not satisfied, the scenario is skipped.

For the TestNtpDiagnostics scenario, it turns out it's possible to
amend it a bit to allow for running in the environments referred
above.

Change-Id: Id5cee9f4a32c18c50fe54c01c7ed81c7e93c5fa7
Reviewed-on: http://gerrit.cloudera.org:8080/15551
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>