Add functionality to serialize a RowBlock into columnar format
This adds the core functionality to take a RowBlock and convert it into
a set of buffers for each column: the cell data, the null bitmap, and
the indirect (string) data.

This also updates wire_protocol-test to add coverage for nulls and
unselected rows, and adds a comparison benchmark. The columnar code path
is 4-5x faster in the best case, and in the worst case only 30% slower
that the existing row-wise code path.

Some follow-up commits will add further optimizations.

Change-Id: I287a8aa6736f19816b0edbe16409c01f35c0319e
Reviewed-on: http://gerrit.cloudera.org:8080/15560
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins